Add files via upload
We have used HTML, CSS, JavaScript & NodeJs for the development of this weather application. In this Project we used OpenWeather Weather API and some of the inbuilt modules of NodeJs. 
Here are some basic terms & terminal command that may help you to understand better:
NPM - Node Package manager is used for installing packages/modules in the nodejs project, this one is automatically installed on your system with the node installation.
npm init - for import package.json file where all the dependencies are available
npm install packageName - for installation of any module/package in the project.
Internal CSS - we have used Internal CSS in this project that means no other css file is imported. 
BONUS*
nodemon - nodemon is the global module which helps to automatically restarts the node application while changing in directory are detected.
npm install nodemon -g (used this command to install nodemon globally on the system)